Implement passing JAccessPoint
